Starting some cleanup
feature: Added RendererFlags to the WindowSpecification; Added a
'double angle' member to the Texture2D class, making it possible for the
texture to be rotated and rendered as such.

change: Rect, Coordinate2D, RectComponent and Coordinate2DComponent now
use floats as their member's datatype instead of int; Renderer::DrawTexture
now uses 'SDL_RenderCopyExF' to render textures at subpixel precision
and render rotated textures

+++ Saddle/src/SDL/Coordinate2D.h
@@ -5,9 +5,9 @@
 namespace Saddle {
 
 struct Coordinate2D {
-    int x, y;
+    float x, y;
 
-    Coordinate2D(int x, int y) : x(x), y(y) { }
+    Coordinate2D(float x, float y) : x(x), y(y) { }
     Coordinate2D(const Coordinate2DComponent& component) : x(component.x), y(component.y) { }
 };
