+++ Sandbox/include/MusicDemo.h
@@ -7,7 +7,7 @@ using namespace Saddle;
     
 class MusicBlock : public Entity {
 public:
-    MusicBlock(const std::string& sound_path, int width, int height, int r = 255, int g = 255, int b = 255)
+    MusicBlock(const std::string& sound_path, float width, float height, int r = 255, int g = 255, int b = 255)
         : m_Sound(sound_path), m_Width(width), m_Height(height)
     {
         AddComponent<TextureComponent>();
